Update battle system

+++ src/main/java/com/aetherwars/model/card/CharacterCard.java
@@ -2,7 +2,7 @@ package com.aetherwars.model.card;
 
 public class CharacterCard extends Card {
   private int attack;
-  private int health;
+  private float health;
   private int attackUp;
   private int healthUp;
   private CharacterType characterType;
@@ -12,7 +12,7 @@ public class CharacterCard extends Card {
     this.characterType = CharacterType.OVERWORLD;
   }
 
-  public CharacterCard(int id, String name, String description, String IMAGE_PATH, CharacterType element, int attack, int attackUp, int health, int healthUp) {
+  public CharacterCard(int id, String name, String description, String IMAGE_PATH, CharacterType element, int attack, int attackUp, float health, int healthUp) {
     super(id, name, description, CardType.CHARACTER, IMAGE_PATH);
     this.characterType = element;
     this.attack = attack;
@@ -21,13 +21,9 @@ public class CharacterCard extends Card {
     this.healthUp = healthUp;
   }
 
-
-
   public void setAttack(int attack) { this.attack = attack; }
   
-  public void setHealth(int health) { this.health = health; }
-
-
+  public void setHealth(float health) { this.health = health; }
 
   public CharacterType getCharacterType(){
     return this.characterType;
@@ -35,7 +31,7 @@ public class CharacterCard extends Card {
 
   public int getAttack() { return this.attack; }
 
-  public int getHealth() { return this.health; }
+  public float getHealth() { return this.health; }
 
   public int getAttackUp() { return this.attackUp; }
